better exceptions and rawRead bugfix
bumped dub version to 0.1.3
cgi.d: fixed a bug where if a bad html form was constructed with inputs
missing the name attributes, the webserver will send a content_length of
0. we were always doing a stdin.rawRead() call.  This will throw an
exception when a zero length buffer is used.   So now we are detecting
this and not trying to read from stdin when there is nothing to read.
cgi.d: added a CGIException class and fixed some stubs where we were
silently failing.